+++ js/data-variables.js
@@ -1,0 +1,45 @@
+/**
+ * Constant variables to create 10 random objects with booking data
+ */
+const TITLES = [
+  'Big apartment',
+  'Regular apartment',
+  'Small apartment',
+  'A private house',
+  'Luxury cottage',
+  'Townhouse on the lake',
+];
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const TIMES = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const DESCRIPTIONS = [
+  'Very nice apartments',
+  'Good shelter from bad weather',
+  'Great castle for knights',
+  'Grandmas cottage with pies',
+  'Lounge paradise',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const LOCATION_LAT_MIN = 35.65;
+const LOCATION_LAT_MAX = 35.7;
+const LOCATION_LNG_MIN = 139.7;
+const LOCATION_LNG_MAX = 139.8;
+const SIMILAR_AD_COUNT = 10;
+
+export {
+  TITLES,
+  TYPES,
+  TIMES,
+  FEATURES,
+  DESCRIPTIONS,
+  PHOTOS,
+  LOCATION_LAT_MIN,
+  LOCATION_LAT_MAX,
+  LOCATION_LNG_MIN,
+  LOCATION_LNG_MAX,
+  SIMILAR_AD_COUNT,
+};
